+++ src/test/java/ai/deepcode/javaclient/core/mocks/AnalysisDataBaseMock.java
@@ -1,5 +1,6 @@
 package ai.deepcode.javaclient.core.mocks;
 
+import ai.deepcode.javaclient.DeepCodeRestApi;
 import ai.deepcode.javaclient.core.AnalysisDataBase;
 import ai.deepcode.javaclient.core.DCLoggerBase;
 import ai.deepcode.javaclient.core.DeepCodeParamsBase;
@@ -12,11 +13,12 @@ import java.util.Collection;
 public class AnalysisDataBaseMock extends AnalysisDataBase {
 
   public AnalysisDataBaseMock(
-      @NotNull PlatformDependentUtilsBase platformDependentUtils,
-      @NotNull HashContentUtilsBase hashContentUtils,
-      @NotNull DeepCodeParamsBase deepCodeParams,
-      @NotNull DCLoggerBase dcLogger) {
-    super(platformDependentUtils, hashContentUtils, deepCodeParams, dcLogger);
+    @NotNull PlatformDependentUtilsBase platformDependentUtils,
+    @NotNull HashContentUtilsBase hashContentUtils,
+    @NotNull DeepCodeParamsBase deepCodeParams,
+    @NotNull DCLoggerBase dcLogger,
+    @NotNull DeepCodeRestApi restApi) {
+    super(platformDependentUtils, hashContentUtils, deepCodeParams, dcLogger, restApi);
   }
 
   @Override