+++ src/test/java/ai/deepcode/javaclient/core/mocks/DeepCodeParamsMock.java
@@ -1,12 +1,13 @@
 package ai.deepcode.javaclient.core.mocks;
 
+import ai.deepcode.javaclient.DeepCodeRestApi;
 import ai.deepcode.javaclient.core.DeepCodeParamsBase;
 import org.jetbrains.annotations.NotNull;
 
 public class DeepCodeParamsMock extends DeepCodeParamsBase {
 
-  public DeepCodeParamsMock() {
-    super(true, "", false, false, 1, "", "", "", () -> 1000L);
+  public DeepCodeParamsMock(DeepCodeRestApi restApi) {
+    super(true, "", false, false, 1, "", "", "", () -> 1000L, restApi);
   }
 
   @Override